requirements_tryout.txt
openai
ipykernel
pillow